src/config/settings.py
import os
from dotenv import load_dotenv
import streamlit as st

load_dotenv()


class Settings:    
    # API Keys
    @staticmethod
    def get_google_api_key():
        try:
            return st.secrets["GOOGLE_API_KEY"]
        except:
            return os.getenv("GOOGLE_API_KEY")
    
    @staticmethod
    def get_pinecone_api_key():
        try:
            return st.secrets["PINECONE_API_KEY"]
        except:
            return os.getenv("PINECONE_API_KEY")
    
    @staticmethod
    def get_groq_api_key():
        try:
            return st.secrets["GROQ_API_KEY"]
        except:
            return os.getenv("GROQ_API_KEY")
    

    # Tambahkan ini di dalam class Settings:
    @staticmethod
    def get_cohere_api_key():
        try:
            return st.secrets["COHERE_API_KEY"]
        except:
            return os.getenv("COHERE_API_KEY")
    
    # Vector Store Configuration
    INDEX_NAME = "ordal-filkom"
    
    # Ubah bagian Model Configuration:
    EMBEDDING_MODEL = "embed-multilingual-v3.0" # Ubah nama modelnya
    
    # LLM Configuration with Fallback
    LLM_MODEL = "llama-3.3-70b-versatile"  # Primary model
    LLM_TEMPERATURE = 0.2
    SIMILARITY_TOP_K = 30
    
    # Fallback models (ordered by priority when primary hits rate limit)
    # Format: (model_name, TPM_limit, description, note)
    FALLBACK_MODELS = [
        ("meta-llama/llama-4-scout-17b-16e-instruct", 30000, "Llama 4 Scout", "mid 🙂"),
        ("llama-3.1-8b-instant", 6000, "Llama 3.1 8B", "agak kocaks 😹"),
    ]
    
    @staticmethod
    def get_all_available_models():
        """
        Get list of all available models (primary + fallbacks)
        
        Returns:
            list: List of dicts with model metadata
                - model: model identifier
                - description: human-readable name
                - tpm: tokens per minute limit
                - note: fun description
        """
        models = [
            {
                "model": Settings.LLM_MODEL,
                "description": "Llama 3.3 70B",
                "tpm": "12,000",
                "note": "paling bagus 🔥"
            }
        ]
        
        # Add fallback models
        for model_name, tpm_limit, description, note in Settings.FALLBACK_MODELS:
            models.append({
                "model": model_name,
                "description": description,
                "tpm": f"{tpm_limit:,}",
                "note": note
            })
        
        return models
    
    # Paths
    DATASET_DIR = "dataset"
    
    # UI Configuration
    PAGE_TITLE = "Ordal Filkom - Asisten Akademik"
    PAGE_ICON = "🎓"
    LAYOUT = "centered"
    
    # Chat Configuration
    MAX_RETRIES = 3
    RETRY_WAIT_BASE = 25
    TOP_SOURCES_TO_DISPLAY = 3
    PDF_RENDER_DPI = 120
