requirements.txt
# LlamaIndex & RAG
llama-index
llama-index-llms-google-genai
llama-index-llms-groq
llama-index-embeddings-google-genai
llama-index-vector-stores-pinecone

# PDF Parsing Enhancement (for tables/diagrams)
llama-parse>=0.5.0
nest-asyncio>=1.5.8
PyMuPDF

# Vector Database
pinecone

# AI/ML
google-generativeai

# UI
streamlit
streamlit-pdf-viewer

# Utils
python-dotenv